demo_segmentation/pre-requirements.txt
torch
torchvision
transformers==4.44.2
flash_attn 
timm 
einops 
numpy
pillow
matplotlib
opencv-python
peft